05/04/2024

+++ list/list-gruppa.php
@@ -5,7 +5,7 @@ if (!Helper::can('admin') && !Helper::can('manager') && !Helper::can('teacher'))
     exit();
 }
 $size = 5;
-if (isset ($_GET['page'])) {
+if (isset($_GET['page'])) {
     $page = Helper::clearInt($_GET['page']);
 } else {
     $page = 1;
@@ -22,7 +22,7 @@ require_once '../template/header.php';
             <section class="content-header">
                 <h3>
                     <b>
-                        <?= $message = isset ($_GET['message']) ? Helper::getQuery($_GET['message']) : 'Список групп' ?>
+                        <?= $message = isset($_GET['message']) ? Helper::getQuery($_GET['message']) : 'Список групп' ?>
                     </b>
                 </h3>
                 <ol class="breadcrumb">
@@ -34,10 +34,9 @@ fa-dashboard"></i> Главная</a></li>
                 </ol>
             </section>
             <div class="box-body">
-                <?php if (Helper::can('admin')) { ?>
+                <?php if (!Helper::can('teacher')): ?>
                     <a class="btn btn-success" href="../add/add-gruppa">Добавить группу</a>
-                <?php }
-                ; ?>
+                <?php endif; ?>
             </div>
             <div class="box-body">
                 <?php
@@ -60,8 +59,11 @@ fa-dashboard"></i> Главная</a></li>
                             <?php
                             foreach ($gruppas as $gruppa) {
                                 echo '<tr>';
-                                echo '<td><a href="../view/view-gruppa?id=' . $gruppa->gruppa_id . '">' . $gruppa->name . '</a> ' . '<a href="../add/add-gruppa?id=' . $gruppa->gruppa_id . '"><i class="fa fa-pencil"></i></a> <a href="../delete/delete-gruppa?id=' . $gruppa->gruppa_id . '"><i class="fa fa-times"></i></a></td>';
-
+                                if (!Helper::can('teacher')):
+                                    echo '<td><a href="../view/view-gruppa?id=' . $gruppa->gruppa_id . '">' . $gruppa->name . '</a> ' . '<a href="../add/add-gruppa?id=' . $gruppa->gruppa_id . '"><i class="fa fa-pencil"></i></a> <a href="../delete/delete-gruppa?id=' . $gruppa->gruppa_id . '"><i class="fa fa-times"></i></a></td>';
+                                else:
+                                    echo '<td><p href="../view/view-gruppa?id=' . $gruppa->gruppa_id . '">' . $gruppa->name . '</p> ' . '</td>';
+                                endif;
                                 echo '<td>' . date(
                                     "d.m.Y",
                                     strtotime($gruppa->date_begin)
